refactor: 지하철 노선 인수테스트 요청 메서드 분리
